magento/graphql-ce#683: Add discount information to Cart
- extracted discount cases to separate test class;
- added case for no discount applied;
- return `null` for discount node if amount is 0.
