+++ dev/tests/api-functional/testsuite/Magento/GraphQl/Quote/Customer/CartTotalsTest.php
@@ -154,44 +154,6 @@ class CartTotalsTest extends GraphQlAbstract
     }
 
     /**
-     * @magentoApiDataFixture Magento/Customer/_files/customer.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/cart_rule_discount_no_coupon.php
-     * @magentoApiDataFixture Magento/GraphQl/Catalog/_files/simple_product.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/customer/create_empty_cart.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/add_simple_product.php
-     */
-    public function testGetDiscountInformation()
-    {
-        $maskedQuoteId = $this->getMaskedQuoteIdByReservedOrderId->execute('test_quote');
-        $query = $this->getQuery($maskedQuoteId);
-        $response = $this->graphQlQuery($query, [], '', $this->getHeaderMap());
-
-        $discountResponse = $response['cart']['prices']['discount'];
-        self::assertEquals(-10, $discountResponse['amount']['value']);
-        self::assertEquals(['50% Off for all orders'], $discountResponse['label']);
-    }
-
-    /**
-     * @magentoApiDataFixture Magento/Customer/_files/customer.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/cart_rule_discount_no_coupon.php
-     * @magentoApiDataFixture Magento/GraphQl/Catalog/_files/simple_product.php
-     * @magentoApiDataFixture Magento/Checkout/_files/discount_10percent_generalusers.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/customer/create_empty_cart.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/add_simple_product.php
-     * @magentoApiDataFixture Magento/GraphQl/Quote/_files/apply_coupon.php
-     */
-    public function testGetDiscountInformationWithTwoRulesApplied()
-    {
-        $maskedQuoteId = $this->getMaskedQuoteIdByReservedOrderId->execute('test_quote');
-        $query = $this->getQuery($maskedQuoteId);
-        $response = $this->graphQlQuery($query, [], '', $this->getHeaderMap());
-
-        $discountResponse = $response['cart']['prices']['discount'];
-        self::assertEquals(-11, $discountResponse['amount']['value']);
-        self::assertEquals(['50% Off for all orders', 'Test Coupon for General'], $discountResponse['label']);
-    }
-
-    /**
      * Generates GraphQl query for retrieving cart totals
      *
      * @param string $maskedQuoteId
@@ -220,13 +182,6 @@ class CartTotalsTest extends GraphQlAbstract
         currency
       }
       applied_taxes {
-        label
-        amount {
-          value
-          currency
-        }
-      }
-      discount {
         label
         amount {
           value
